backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.11
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
librt==0.8.1
packaging==26.0
pathspec==1.0.4
psycopg2-binary==2.9.11
PyJWT==2.11.0
sqlparse==0.5.5
tomli==2.4.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
whitenoise==6.8.2
factory-boy==3.3.0
